==> 2-Feature Engineering/1.Feature Engineering.ipynb <==
# Data Preprocessing for Emotional Status Project
Author: Sicong Zhao

## 1.Load Data

import pickle
import numpy as np
import pandas as pd
from functools import reduce
import datetime
from tqdm import tqdm # Show pandas progress

# Experience Sampling data, which contains 8 metrics labels for emotion states.
# We are using these 8 metrics to generate the positive or negative of the emotion, and make the prediction.
with open('data/EXPERIENCING_SAMPLE_R00_DND.pickle', 'rb') as handle:
    exp = pickle.load(handle)

# Predictor: Steps, by minute, from Fitbit
# with open('data/STEPS_R00_DND.pickle', 'rb') as handle:
#     steps = pickle.load(handle)
steps = pd.read_csv('steps_with_datetime.csv')
steps['datetime'] = pd.to_datetime(steps['datetime'], errors='coerce')

# Predictor: Heart Rate, by minute, from Fitbit
# with open('data/HR_R00_DND.pickle', 'rb') as handle:
#     hr = pickle.load(handle)
hr = pd.read_csv('hr_with_datetime.csv')
hr['datetime'] = pd.to_datetime(hr['datetime'], errors='coerce')

# Predictor: Sleep, by day, from Fitbit
with open('data/SLEEP_R00_DND.pickle', 'rb') as handle:
    sleep = pickle.load(handle)

# Predicor: Survey, contains demographic, health, personality data, one time measurement for each subject.
with open('data/SURVEY_R00_DND.pickle', 'rb') as handle:
    survey = pickle.load(handle)

# A lot of missing data in survey
survey['AVI'].isna().sum() / survey['AVI'].shape[0]

## 2.Pre-processing

### (1) Convert ID to string

steps['ID'] = steps['ID'].astype('str')
sleep['ID'] = sleep['ID'].astype('str')
hr['ID'] = hr['ID'].astype('str')

### (2) Add timestamp for heart rate, step data

# hr['datetime'] = hr['Date'].astype(str) + ' ' + hr['TIME'].astype(str)
# hr['datetime'] = pd.to_datetime(hr['datetime'], errors='coerce')
# hr.to_csv('hr_with_datetime.csv')

# steps['datetime'] = steps['Date'].astype(str) + ' ' + steps['TIME'].astype(str)
# steps['datetime'] = pd.to_datetime(steps['datetime'], errors='coerce')
# steps.to_csv('steps_with_datetime.csv')

### (3) Extract useful features and merge survey data

survey.keys()

# Get columns for Medical Screening
med_df = survey['MEDICAL_SCREENING'][['Subject','BMI']]

# Get columns for SBQ
sbq_df = survey['SBQ'][['Subject', 'SBQ']]

# Get columns for FTP
ftp_df = survey['FTP'][['Subject', 'FTP']]

# Get columns for SWLS
swls_df = survey['SWLS'][['Subject', 'SWLS']]

# Get columns for NEO-SF
neosf_df = survey['NEO_SF'][['Subject', 'Neuroticism', 'Extraversion', 'Openness', 'Agreeableness','Conscientiousness']]

# Get columns for TPQ-NS
tpqns_df = survey['TPQ_NS'][['Subject','NS_total']]

# Get columns for BIS
bis_df = survey['BIS'][['Subject','BIS_total']]

# Get columns for BISBAS
bisbas_cols = ['Subject', 'BIS.5', 'BAS_D', 'BAS_FS', 'BAS_RR']
bisbas_df = survey['BISBAS'][bisbas_cols]

# Get columns for AVI
avi_col_actual = [x for x in survey['AVI'].columns.values if 'actual' in x]
avi_col_ideal = [x for x in survey['AVI'].columns.values if 'ideal' in x]
avi_col = np.concatenate([['Subject'], avi_col_actual,avi_col_ideal])
avi_df = survey['AVI'][avi_col]
avi_df['Subject'] = avi_df['Subject'].astype('str')

# Get columns for Demographics
demo_df = survey['DEMOGRAPHICS'][['Subject', 'Education', 'Ethnicity', 'Sex', 'Marital_Status', 'Children', 'Household_income', 'Religion', 'Age', 'Medications']]

# Merge all dataframes, which contains the features for modeling
data_frames = [sbq_df, ftp_df, swls_df, neosf_df, tpqns_df, bis_df, bisbas_df, avi_df, demo_df, med_df]
for df in data_frames:
    df['Subject'] = df['Subject'].astype(str)

survey_df = reduce(lambda  left,right: pd.merge(left,right,on=['Subject'],
                                            how='left'), data_frames)

# Convert data type
survey_df['Ethnicity'] = survey_df['Ethnicity'].astype(str, errors = 'raise')
survey_df['Sex'] = survey_df['Sex'].astype(str, errors = 'raise')
survey_df['Marital_Status'] = survey_df['Marital_Status'].astype(str, errors = 'raise')
survey_df['Household_income'] = survey_df['Household_income'].astype(str, errors = 'raise')
survey_df['Religion'] = survey_df['Religion'].astype(str, errors = 'raise')
survey_df['Medications'] = survey_df['Medications'].astype(str, errors = 'raise')

# Missing data detection
survey_df.isna().sum()/survey_df.shape[0]

survey_df.head()

survey_df.dtypes

(6) Fix ID inconsistency

# Fix subject ID inconsistant issue, only run once
exp['subject'] = exp['subject'].str.replace('DND', 'DND1', regex=False)
survey_df['Subject'] = survey_df['Subject'].str.replace('DND', 'DND1', regex=False)

## 3.Feature Engineering

### (1) Time window for feature engineering

# Generate 30 mins prior time point
exp['start_survey_5m_ahead'] = exp['start_survey'] - datetime.timedelta(minutes=5)
exp['start_survey_10m_ahead'] = exp['start_survey'] - datetime.timedelta(minutes=10)
exp['start_survey_30m_ahead'] = exp['start_survey'] - datetime.timedelta(minutes=30)
exp['start_survey_1h_ahead'] = exp['start_survey'] - datetime.timedelta(minutes=60)
exp['start_survey_3h_ahead'] = exp['start_survey'] - datetime.timedelta(hours=3)

### (2) Step related feature

Steps related features:
* Statistics: max, min, mean, std
* Move Rate: The # of minutes with step > 0 / total minutes
* Active Rate: The # of minutes with step > 10 / total minutes
* Very Active Rate: The # of minutes with step > 20 / total minutes
* Running Rate: The # of minutes with step > 30 / total minutes

def step_feature_gen(df):
    if df.shape[0] == 0:
        return [None,None,None,None,None,None,None,None,None]
    else:
        max_step = df['VALUE'].max()
        min_step = df['VALUE'].min()
        mean = df['VALUE'].mean()
        var = df['VALUE'].var()
        median = df['VALUE'].median()
        move_rate = df[df['VALUE'] > 0].shape[0]/df.shape[0]
        active_rate = df[df['VALUE'] > 10].shape[0]/df.shape[0]
        very_active_rate = df[df['VALUE'] > 20].shape[0]/df.shape[0]
        running_rate = df[df['VALUE'] > 30].shape[0]/df.shape[0]
        return [max_step, min_step, mean, var, median, move_rate, active_rate, very_active_rate, running_rate]

def gen_step_features_by_id(subject_id, index, row, step_by_id):
    start_3h_time = row['start_survey_3h_ahead']
    start_1h_time = row['start_survey_1h_ahead']
    start_30m_time = row['start_survey_30m_ahead']
    start_10m_time = row['start_survey_10m_ahead']
    start_5m_time = row['start_survey_5m_ahead']
    end_time = row['start_survey']
    df_before_exp = step_by_id[step_by_id['datetime'] <= end_time]
    df_3h = df_before_exp[df_before_exp['datetime'] >= start_3h_time]
    df_1h = df_before_exp[df_before_exp['datetime'] >= start_1h_time]
    df_30m = df_before_exp[df_before_exp['datetime'] >= start_30m_time]
    df_10m = df_before_exp[df_before_exp['datetime'] >= start_10m_time]
    df_5m = df_before_exp[df_before_exp['datetime'] >= start_5m_time]
    # Generate features
    step_features_3h = step_feature_gen(df_3h)
    step_features_1h = step_feature_gen(df_1h)
    step_features_30m = step_feature_gen(df_30m)
    step_features_10m = step_feature_gen(df_10m)
    step_features_5m = step_feature_gen(df_5m)
    prev_max_step = df_before_exp['VALUE'].max()
    prev_min_step = df_before_exp['VALUE'].min()
    prev_med_step = df_before_exp['VALUE'].median()
    return [index, subject_id, prev_max_step, prev_min_step, prev_med_step] \
            + step_features_3h + step_features_1h + step_features_30m + step_features_10m + step_features_5m

Step1. Get subject id

steps_ids = np.unique(steps['ID'].values)
exp_ids = np.unique(exp.subject.values)
# Detect subject without experience sampling data
print(set(steps_ids) - set(exp_ids))
# Detect subject without step data
print(set(exp_ids) - set(steps_ids))

Step2. Calculate step features for all subject

# Get All Unique IDs, remove subject without step data. (we could not construct step features for them.)
exp_ids = np.unique(exp.subject.values)
exp_id_1 = np.in1d(exp_ids, 'DND1033').nonzero()[0]
exp_id_2 = np.in1d(exp_ids, '1079').nonzero()[0]
exp_id_3 = np.in1d(exp_ids, '1097').nonzero()[0]

exp_ids = np.delete(exp_ids, [exp_id_1,exp_id_2,exp_id_3])
    
# Result stores here
step_features = []
# Count for progress
count = 0
# Iterate each id, decrease redundant computation
for exp_id in exp_ids:
    # Progress output
    count += 1
    b = str(count) + '/'+ str(len(exp_ids))
    print (b, end="\r")
    # Get subject_id
    subject_id = exp_id
    # Extract exp dataframe by subject_id
    exp_by_id = exp[exp['subject'].astype(str) == subject_id]
    # Extract heart rate dataframe by subject_id
    step_by_id = steps[steps['ID'].astype(str) == subject_id]
    # Generate features for each experience sampling of each subject
    for index, row in exp_by_id.iterrows():
        feature_by_row = gen_step_features_by_id(exp_id, index, row, step_by_id)
        step_features.append(feature_by_row)

Step3. Convert results to dataframe, and drop NA

# steps_features = pd.DataFrame(steps_features.dropna())
steps_features_colnames = ['index', 'subject_id', 'step_max', 'step_min', 'step_median',
                'steps_max_3h','steps_min_3h','steps_mean_3h', 'steps_var_3h', 'steps_median_3h', 'move_rate_3h', 'active_rate_3h', 'very_active_rate_3h', 'running_rate_3h',
                'steps_max_1h','steps_min_1h','steps_mean_1h', 'steps_var_1h', 'steps_median_1h', 'move_rate_1h', 'active_rate_1h', 'very_active_rate_1h', 'running_rate_1h',
                'steps_max_30m','steps_min_30m','steps_mean_30m', 'steps_var_30m', 'steps_median_30m', 'move_rate_30m', 'active_rate_30m', 'very_active_rate_30m', 'running_rate_30m',
                'steps_max_10m','steps_min_10m','steps_mean_10m', 'steps_var_10m', 'steps_median_10m', 'move_rate_10m', 'active_rate_10m', 'very_active_rate_10m', 'running_rate_10m',
                'steps_max_5m','steps_min_5m','steps_mean_5m', 'steps_var_5m', 'steps_median_5m', 'move_rate_5m', 'active_rate_5m', 'very_active_rate_5m', 'running_rate_5m']
step_features_df = pd.DataFrame(step_features)
step_features_df.columns = steps_features_colnames
# Drop NA
step_features_cl = step_features_df.dropna()
step_features_cl.to_csv('step_features_clean.csv')

### (3) Heart Rate related features

Heart Rate related features:
* Statistics: max, min, mean, std
* Resting Rate: The # of minutes with HR < 30 percentile heart rate > 0 / total minutes
* Moderate Rate: The # of minutes with HR > 50 percentile heart rate > 0 / total minutes
* Very Active Rate: The # of minutes with HR > 80 percentile heart rate > 0 / total minutes
* SDNN: Standard deviation of heartbeat intervals
* pHR2: Percentage of the difference between ajacent HR greater than 2
* rMSSD: Root of mean squared HR change
* Highest HR
* Lowest HR
* l_h: Lowest HR / Highest HR
* CR: Highest HR / Highest HR so far

def hr_feature_gen(df, rest_thred, moderate_thred, active_thred, hr_max):
    if df.shape[0] == 0:
        return [None, None, None, None, None, None, None, None, None, None, None, None, None, None]
    else:
        mean = df['VALUE'].mean()
        var = df['VALUE'].var()
        std = df['VALUE'].std()
        median = df['VALUE'].median()
        rest_rate = df[df['VALUE'] <= rest_thred].shape[0]/df.shape[0]
        moderate_rate = df[df['VALUE'] > moderate_thred].shape[0]/df.shape[0]
        very_active_rate = df[df['VALUE'] > active_thred].shape[0]/df.shape[0]
        
        hr_interval_all = np.array([])
        for index, row in df.iterrows():
            hr_by_min = row['VALUE']
            hr_interval_min = np.full((1, hr_by_min), 1/hr_by_min)[0]
            hr_interval_all = np.concatenate([hr_interval_min, hr_interval_all])
        # SDNN - Standard deviation of heartbeat intervals
        SDNN = hr_interval_all.std()
        # pHR2 - percentage of the difference between ajacent HR greater than 2
        hr_seq = df['VALUE'].values
        hr_diff = abs(hr_seq[:-1] - hr_seq[1:])
        # rMSSD
        if len(hr_diff) == 0:
            pHR2 = None
            rMSSD = None
        else:
            pHR2 = sum(hr_diff > 2)/len(hr_diff)
            rMSSD = np.sqrt(sum(hr_diff**2)/len(hr_diff))
            # Lowest heart rate
        low_hr = df['VALUE'].min()
        # Highest heart rate
        high_hr = df['VALUE'].max()
        # LR/HR
        if high_hr == 0:
            l_h = None
        else:
            l_h = low_hr/high_hr
        # CR - high_hr/hr_max_of_all_time
        if hr_max == 0:
            CR = None
        else:
            CR = high_hr/hr_max
        
        return [SDNN, pHR2, rMSSD, low_hr, high_hr, l_h, CR, mean, var, std, median, rest_rate, moderate_rate, very_active_rate]

def gen_hr_features_by_id(subject_id, index, row, hr_by_id):
    start_3h_time = row['start_survey_3h_ahead']
    start_1h_time = row['start_survey_1h_ahead']
    start_30m_time = row['start_survey_30m_ahead']
    start_10m_time = row['start_survey_10m_ahead']
    start_5m_time = row['start_survey_5m_ahead']
    end_time = row['start_survey']
    
    df_before_exp = hr_by_id[hr_by_id['datetime'] <= end_time]
    hr_percentile = df_before_exp['VALUE'].quantile([0, .3, .5, .8, 1]).values
    rest_thred = hr_percentile[1]
    moderate_thred = hr_percentile[2]
    active_thred = hr_percentile[3]

    df_3h = df_before_exp[df_before_exp['datetime'] >= start_3h_time]
    df_1h = df_before_exp[df_before_exp['datetime'] >= start_1h_time]
    df_30m = df_before_exp[df_before_exp['datetime'] >= start_30m_time]
    df_10m = df_before_exp[df_before_exp['datetime'] >= start_10m_time]
    df_5m = df_before_exp[df_before_exp['datetime'] >= start_5m_time]
    # All time features
    hr_max = df_before_exp['VALUE'].max()
    hr_min = df_before_exp['VALUE'].min()
    hr_med = df_before_exp['VALUE'].median()
    # Generate features
    hr_features_3h = hr_feature_gen(df_3h, rest_thred, moderate_thred, active_thred, hr_max)
    hr_features_1h = hr_feature_gen(df_1h, rest_thred, moderate_thred, active_thred, hr_max)
    hr_features_30m = hr_feature_gen(df_30m, rest_thred, moderate_thred, active_thred, hr_max)
    hr_features_10m = hr_feature_gen(df_10m, rest_thred, moderate_thred, active_thred, hr_max)
    hr_features_5m = hr_feature_gen(df_5m, rest_thred, moderate_thred, active_thred, hr_max)
    
    return [index, subject_id, hr_max, hr_min, hr_med] + [hr_percentile[0],hr_percentile[1],hr_percentile[2],hr_percentile[3],hr_percentile[4]]\
            + hr_features_3h + hr_features_1h + hr_features_30m + hr_features_10m + hr_features_5m

# Detect ids in hr but not in exp
hr_ids = np.unique(hr['ID'].values)
exp_ids = np.unique(exp.subject.values)
print(set(hr_ids) - set(exp_ids))
# Detect ids in exp but not in hr
print(set(exp_ids) - set(hr_ids))

Step1. Get Subject ID

# Get All Unique IDs
exp_ids = np.unique(exp.subject.values)
exp_id_1 = np.in1d(exp_ids, 'DND1033').nonzero()[0]
exp_ids = np.delete(exp_ids, exp_id_1)

Step2. Calculate features by subject.

# Count for progress
count = 0
# Result stores here
hr_features = []
# Iterate each id, decrease redundant computation
for exp_id in exp_ids:
    # Progress output
    count += 1
    b = str(count) + '/'+ str(len(exp_ids))
    print (b, end="\r")
    # Get subject_id
    subject_id = exp_id
    # Extract exp dataframe by subject_id
    exp_by_id = exp[exp['subject'].astype(str) == subject_id]
    # Extract heart rate dataframe by subject_id
    hr_by_id = hr[(hr['ID'].astype(str) == subject_id)]
    # Generate features for each experience sampling of each subject
    for index, row in exp_by_id.iterrows():
        feature_by_row = gen_hr_features_by_id(exp_id, index, row, hr_by_id)
        hr_features.append(feature_by_row)

Step3. Convert results to dataframe, and drop NA

hr_features_colnames = ['index', 'subject_id', 'hr_max', 'hr_min', 'hr_med', 'hr_0','hr_0.3','hr_0.5','hr_0.8','hr_1',
                         'SDNN_3h', 'pHR2_3h', 'rMSSD_3h', 'low_hr_3h', 'high_hr_3h', 'l_h_3h', 'CR_3h',
                         'hr_mean_3h', 'hr_var_3h', 'hr_std_3h', 'hr_median_3h', 'hr_rest_rate_3h', 'hr_moderate_rate_3h', 'hr_very_active_rate_3h',
                         'SDNN_1h', 'pHR2_1h', 'rMSSD_1h', 'low_hr_1h', 'high_hr_1h', 'l_h_1h', 'CR_1h',
                         'hr_mean_1h', 'hr_var_1h', 'hr_std_1h', 'hr_median_1h', 'hr_rest_rate_1h', 'hr_moderate_rate_1h', 'hr_very_active_rate_1h',
                         'SDNN_30m', 'pHR2_30m', 'rMSSD_30m', 'low_hr_30m', 'high_hr_30m', 'l_h_30m', 'CR_30m',
                         'hr_mean_30m', 'hr_var_30m', 'hr_std_30m', 'hr_median_30m', 'hr_rest_rate_30m', 'hr_moderate_rate_30m', 'hr_very_active_rate_30m',
                         'SDNN_10m', 'pHR2_10m', 'rMSSD_10m', 'low_hr_10m', 'high_hr_10m', 'l_h_10m', 'CR_10m',
                         'hr_mean_10m', 'hr_var_10m', 'hr_std_10m', 'hr_median_10m', 'hr_rest_rate_10m', 'hr_moderate_rate_10m', 'hr_very_active_rate_10m',
                         'SDNN_5m', 'pHR2_5m', 'rMSSD_5m', 'low_hr_5m', 'high_hr_5m', 'l_h_5m', 'CR_5m',
                         'hr_mean_5m', 'hr_var_5m', 'hr_std_5m', 'hr_median_5m', 'hr_rest_rate_5m', 'hr_moderate_rate_5m', 'hr_very_active_rate_5m']
hr_features_df = pd.DataFrame(hr_features)
hr_features_df.columns = hr_features_colnames
# Drop NA
hr_features_cl = hr_features_df.dropna()
# Export
hr_features_cl.to_csv('hr_features_clean.csv')

pd.DataFrame(exp_steps_hr.isna().sum() / exp_steps_hr.shape[0])

## 4. Merge Results

# Reset index for merging
step_features_cl = step_features_cl.set_index('index')
hr_features_cl = hr_features_cl.set_index('index')
# Convert id column data type for merging
exp_steps_hr['subject'] = exp_steps_hr['subject'].astype(str)
sleep['ID'] = sleep['ID'].astype(str)
# Create 'survey_date' in exp for merging with sleep
exp['start_survey'] = pd.to_datetime(exp['start_survey'])
exp['survey_date'] = pd.to_datetime(exp['start_survey'].dt.date)
# Add 'survey_date' to sleep data
sleep['Date'] = pd.to_datetime(sleep['Date'])
sleep['Date'] = pd.to_datetime(sleep['Date'].dt.date)

# Merge design matrix progressively
exp_steps = pd.merge(exp, step_features_cl, how='left', left_index=True, right_index=True)
exp_steps_hr = pd.merge(exp_steps, hr_features_cl, how='left', left_index=True, right_index=True)
exp_steps_hr_sleep = pd.merge(exp_steps_hr, sleep, how='left', left_on=['subject','survey_date'], right_on=['ID','Date'])

exp_steps_hr_sleep.to_csv('exp_steps_hr_sleep.csv')

exp_steps_hr_sleep_survey = pd.merge(exp_steps_hr_sleep, survey_df, how='left', left_on=['subject'], right_on=['Subject'])

# Add survey hour as a feature
exp_steps_hr_sleep_survey['survey_hour'] = exp_steps_hr_sleep_survey.start_survey.dt.hour
# Export data
exp_steps_hr_sleep_survey.to_csv('exp_steps_hr_sleep_survey.csv')

# Missing data detection
missing_df = exp_steps_hr_sleep_survey.isna().sum()/exp_steps_hr_sleep_survey.shape[0]
# Export missing data
pd.DataFrame(missing_df).to_csv('missingness.csv')
